app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "azure-identity>=1.23.1",
    "azure-search-documents>=11.5.3",
    "azure-storage-blob>=12.26.0",
    "gunicorn>=23.0.0",
    "pdfplumber>=0.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.1.0",
]
